@@ -1,23 +1,24 @@
 require 'base64'
 
 module Admin; end
+
 class Admin::ContentController < Admin::BaseController
-  layout "administration", :except => [:show, :autosave]
+  layout "administration", except: [:show, :autosave]
 
   cache_sweeper :blog_sweeper
 
   def auto_complete_for_article_keywords
     @items = Tag.find_with_char params[:article][:keywords].strip
-    render :inline => "<%= raw auto_complete_result @items, 'name' %>"
+    render inline: "<%= raw auto_complete_result @items, 'name' %>"
   end
 
   def index
     @search = params[:search] ? params[:search] : {}
-    
-    @articles = Article.search_with_pagination(@search, {:page => params[:page], :per_page => this_blog.admin_display_elements})
+
+    @articles = Article.search_with_pagination(@search, {page: params[:page], per_page: this_blog.admin_display_elements})
 
     if request.xhr?
-      render :partial => 'article_list', :locals => { :articles => @articles }
+      render partial: 'article_list', locals: { articles: @articles }
     else
       @article = Article.new(params[:article])
     end
@@ -30,7 +31,7 @@ class Admin::ContentController < Admin::BaseController
   def edit
     @article = Article.find(params[:id])
     unless @article.access_by? current_user
-      redirect_to :action => 'index'
+      redirect_to action: 'index'
       flash[:error] = _("Error, you are not allowed to perform this action")
       return
     end
@@ -44,7 +45,7 @@ class Admin::ContentController < Admin::BaseController
       flash[:error] = _("Error, you are not allowed to perform this action")
       return(redirect_to :action => 'index')
     end
-    
+
     return(render 'admin/shared/destroy') unless request.post?
 
     @record.destroy
@@ -52,18 +53,23 @@ class Admin::ContentController < Admin::BaseController
     redirect_to :action => 'index'
   end
 
-  def insert_editor
-    editor = 'visual'
-    editor = 'simple' if params[:editor].to_s == 'simple'
-    current_user.editor = editor
-    current_user.save!
+  def resource_do(action)
+    @article = Article.find(params[:id])
+    @resources = Resource.by_created_at
+    @resource = Resource.find(params["resource_id"])
+    @article.resources.send(action, resource)
+    @article.save
 
-    render :partial => "#{editor}_editor"
+    render :partial => "show_resources"
   end
 
-  def category_add; do_add_or_remove_fu; end
-  alias_method :resource_add,    :category_add
-  alias_method :resource_remove, :category_add
+  def resource_add
+    resource_do(:<<)
+  end
+
+  def resource_remove
+    resource_do(:delete)
+  end
 
   def attachment_box_add
     render :update do |page|
@@ -75,32 +81,26 @@ class Admin::ContentController < Admin::BaseController
     end
   end
 
-  def attachment_save(attachment)
-    begin
-      Resource.create(:filename => attachment.original_filename, :mime => attachment.content_type.chomp, 
-                      :created_at => Time.now).write_to_disk(attachment)
-    rescue => e
-      logger.info(e.message)
-      nil
-    end
-  end
-
   def autosave
     id = params[:id]
     id = params[:article][:id] if params[:article] && params[:article][:id]
+
     @article = Article.get_or_build_article(id)
     @article.text_filter = current_user.text_filter if current_user.simple_editor?
 
     get_fresh_or_existing_draft_for_article
-    
+
     @article.attributes = params[:article]
     @article.published = false
-    set_article_author
-    save_attachments
-
-    set_article_title_for_autosave
-
+    @article.set_author(current_user)
+    @article.save_attachments!(params[:attachments])
     @article.state = "draft" unless @article.state == "withdrawn"
+
+    if @article.title.blank?
+      lastid = Article.find(:first, :order => 'id DESC').id
+      @article.title = "Draft article " + lastid.to_s
+    end
+
     if @article.save
       render(:update) do |page|
         page.replace_html('autosave', hidden_field_tag('article[id]', @article.id))
@@ -127,18 +127,6 @@ class Admin::ContentController < Admin::BaseController
 
   attr_accessor :resources, :categories, :resource, :category
 
-  def do_add_or_remove_fu
-    attrib, action = params[:action].split('_')
-    @article = Article.find(params[:id])
-    self.send("#{attrib}=", self.class.const_get(attrib.classify).find(params["#{attrib}_id"]))
-    send("setup_#{attrib.pluralize}")
-    @article.send(attrib.pluralize).send(real_action_for(action), send(attrib))
-    @article.save
-    render :partial => "show_#{attrib.pluralize}"
-  end
-
-  def real_action_for(action); { 'add' => :<<, 'remove' => :delete}[action]; end
-
   def new_or_edit
     id = params[:id]
     id = params[:article][:id] if params[:article] && params[:article][:id]
@@ -159,18 +147,25 @@ class Admin::ContentController < Admin::BaseController
     @article.keywords = Tag.collection_to_string @article.tags
     @article.attributes = params[:article]
     # TODO: Consider refactoring, because double rescue looks... weird.
-        
+
     @article.published_at = DateTime.strptime(params[:article][:published_at], "%B %e, %Y %I:%M %p GMT%z").utc rescue Time.parse(params[:article][:published_at]).utc rescue nil
 
     if request.post?
-      set_article_author
-      save_attachments
-      
+      @article.set_author(current_user)
+
+      @article.save_attachments!(params[:attachments])
       @article.state = "draft" if @article.draft
 
       if @article.save
-        destroy_the_draft unless @article.draft
-        set_article_categories
+        unless @article.draft
+          Article.all(conditions: { parent_id: @article.id }).map(&:destroy)
+        end
+        @article.categorizations.clear
+        if params[:categories]
+          Category.find(params[:categories]).each do |cat|
+            @article.categories << cat
+          end
+        end
         set_the_flash
         redirect_to :action => 'index'
         return
@@ -194,50 +189,4 @@ class Admin::ContentController < Admin::BaseController
     end
   end
 
-  def destroy_the_draft
-    Article.all(:conditions => { :parent_id => @article.id }).map(&:destroy)
-  end
-
-  def set_article_author
-    return if @article.author
-    @article.author = current_user.login
-    @article.user   = current_user
-  end
-
-  def set_article_title_for_autosave
-    if @article.title.blank?
-      lastid = Article.find(:first, :order => 'id DESC').id
-      @article.title = "Draft article " + lastid.to_s
-    end
-  end
-
-  def save_attachments
-    return if params[:attachments].nil?
-    params[:attachments].each do |k,v|
-      a = attachment_save(v)
-      @article.resources << a unless a.nil?
-    end
-  end
-
-  def set_article_categories
-    @article.categorizations.clear
-    if params[:categories]
-      Category.find(params[:categories]).each do |cat|
-        @article.categories << cat
-      end
-    end
-  end
-
-  def def_build_body
-    if @article.body =~ /<!--more-->/
-      body = @article.body.split('<!--more-->')
-      @article.body = body[0]
-      @article.extended = body[1]
-    end
-
-  end
-
-  def setup_resources
-    @resources = Resource.by_created_at
-  end
 end
