@@ -2,8 +2,9 @@ module Admin::ContentHelper
   def link_to_destroy_draft(record, controller = controller.controller_name)
     return unless record.state.to_s.downcase == "draft"
     link_to(_("Destroy this draft"),
-      { :controller => controller, :action => 'destroy', :id => record.id },
-        :confirm => _("Are you sure?"), :method => :post, :class => 'btn danger' )
+      { controller: controller, action: 'destroy', id: record.id },
+        data: { confirm: _("Are you sure?")},
+        method: :post, class: 'btn danger' )
   end
 
   def auto_complete_result(entries, field, phrase = nil)
@@ -36,7 +37,7 @@ module Admin::ContentHelper
 
     javascript_tag(function)
   end
-  
+
   def get_feedback_status(status)
     status ? _("enabled") : _("disabled")
   end
@@ -46,19 +47,19 @@ module Admin::ContentHelper
     if post_types.size.zero?
       return hidden_field_tag "article[post_type]", "read"
     end
-    
+
     html = "<div class='well'>"
     html << content_tag(:h4, _("Article type"))
     html << "<select name=article[post_type]>"
-        
+
     post_types.each do |pt|
       html << "<option value='read' #{'selected' if @article.post_type == 'read'} >#{_('Default')}</option>"
       html << "<option #{'selected' if @article.post_type == pt.permalink} value='#{pt.permalink}'>#{pt.name}</option>"
     end
-    
+
     html << "</select>"
     html << "</div>"
-  end  
+  end
   
   def toggle_element(element)
     "<a href=\"#\" onclick=\"Element.toggle('#{element}'); return false\">#{_("Change")}</a>".html_safe
