@@ -7,13 +7,19 @@ class SplitsContentRulesArticlesPages < ActiveRecord::Migration
 
   def self.up
     say "Splits content modules into pages and articles"
-    Profile.find_by_label("admin").update_attributes(:modules => [:dashboard, :articles, :pages, :media, :feedback, :themes, :sidebar, :users, :settings, :profile, :seo])
-    Profile.find_by_label("publisher").update_attributes(:modules => [:dashboard, :articles, :media, :pages, :feedback, :profile])
+
+    Profile.find_by_label("admin").try :update_attributes,
+      modules: [:dashboard, :articles, :pages, :media, :feedback, :themes, :sidebar, :users, :settings, :profile, :seo]
+    Profile.find_by_label("publisher").try :update_attributes,
+      modules: [:dashboard, :articles, :media, :pages, :feedback, :profile]
   end
 
   def self.down
     say "Merges pages and articles to content"
-    Profile.find_by_label("admin").update_attributes(:modules => [:dashboard, :write, :content, :feedback, :themes, :sidebar, :users, :settings, :profile, :seo ])
-    Profile.find_by_label("publisher").update_attributes(:modules => [:dashboard, :write, :content, :feedback, :profile ])
+
+    Profile.find_by_label("admin").try :update_attributes,
+      modules: [:dashboard, :write, :content, :feedback, :themes, :sidebar, :users, :settings, :profile, :seo]
+    Profile.find_by_label("publisher").try :update_attributes,
+      modules: [:dashboard, :write, :content, :feedback, :profile]
   end
 end
