@@ -2,7 +2,7 @@
 require 'spec_helper'
 
 describe ApplicationHelper do
-  before(:each) { @blog = FactoryGirl.create(:blog) }
+  before(:each) { @blog = build_stubbed(:blog) }
   describe '#render_flash' do
     it 'should render empty string if no flash' do
       render_flash.should == ''
@@ -14,30 +14,32 @@ describe ApplicationHelper do
     end
 
     it 'should render the notice and error flash' do
-      flash[:notice] = 'good update'
-      flash[:error] = "it's not good"
+      flash[:notice] = 'good update'.html_safe
+      flash[:error] = "it's not good".html_safe
       render_flash.split("<br />\n").sort.should == ['<span class="error">it\'s not good</span>','<span class="notice">good update</span>']
     end
   end
 
   describe "#link_to_permalink" do
     describe "for a simple ascii-only permalink" do
-      subject { link_to_permalink(FactoryGirl.create(:article, :published_at => Date.new(2004, 6, 1)), "title") }
-      it { should be_html_safe }
+      subject { link_to_permalink(
+        build_stubbed(:article, published_at: Date.new(2004, 6, 1).to_datetime),
+        "title") }
+      it { subject.html_safe?.should be_true }
       it { should == '<a href="http://myblog.net/2004/06/01/a-big-article">title</a>' }
     end
 
     describe "for a multibyte permalink" do
-      subject { FactoryGirl.create(:article, :permalink => 'ルビー') }
+      subject { build_stubbed(:article, :permalink => 'ルビー') }
       it "escapes the multibyte characters" do
         link_to_permalink(subject, "title").should =~ /%E3%83%AB%E3%83%93%E3%83%BC/
       end
     end
   end
-  
+
   describe '#display_date' do
-    before(:each) { 
-      @article = FactoryGirl.create(:article, :body => "hello world and im herer") 
+    before(:each) {
+      @article = build_stubbed(:article, :body => "hello world and im herer")
     }
 
     ['%d/%m/%y', '%m/%m/%y', '%d %b %Y', '%b %d %Y'].each do |spec|
@@ -46,13 +48,40 @@ describe ApplicationHelper do
         display_date(@article.published_at).should == @article.published_at.strftime(spec)
       end
     end
-        
+
     ['%I:%M%p', '%H:%M', '%Hh%M'].each do |spec|
       it "should return time with format #{spec}" do
         @blog.time_format = spec
         display_time(@article.published_at).should == @article.published_at.strftime(spec)
       end
-    end    
+    end
   end
-  
+
+  describe "stop_index_robots?" do
+    it "returns false by default" do
+      helper.stop_index_robots?.should be_false
+    end
+
+    it "returns true when params[:year] present" do
+      params[:year] = 2010
+      helper.stop_index_robots?.should be_true
+    end
+
+    it "returns true when params[:page] present" do
+      params[:page] = 2
+      helper.stop_index_robots?.should be_true
+    end
+
+    it "returns true when controller is tags and blog configure to unindex tags page" do
+      Blog.any_instance.should_receive(:unindex_tags).and_return(true)
+      helper.stub(:controller_name).and_return("tags")
+      helper.stop_index_robots?.should be_true
+    end
+
+    it "returns true when controller is tags and blog configure to unindex tags page" do
+      Blog.any_instance.should_receive(:unindex_categories).and_return(true)
+      helper.stub(:controller_name).and_return("categories")
+      helper.stop_index_robots?.should be_true
+    end
+  end
 end
