@@ -34,8 +34,10 @@ class Admin::SettingsController < Admin::BaseController
   def update_database
     @current_version = Migrator.current_schema_version
     @needed_version = Migrator.max_schema_version
+    puts "Current: #{@current_version}"
+    puts "Needed: #{@needed_version}"
     @support = Migrator.db_supports_migrations?
-    @needed_migrations = Migrator.available_migrations[@current_version..@needed_version].collect do |mig|
+    @needed_migrations = Migrator.needed_migrations(@current_version).collect do |mig|
       mig.scan(/\d+\_([\w_]+)\.rb$/).flatten.first.humanize
     end
   end
