@@ -1,5 +1,5 @@
 class ArticlesController < ContentController
-  before_filter :login_required, :only => [:preview]
+  before_filter :login_required, :only => [:preview, :preview_page]
   before_filter :auto_discovery_feed, :only => [:show, :index]
   before_filter :verify_config
 
@@ -20,16 +20,24 @@ class ArticlesController < ContentController
     end
 
     unless params[:year].blank?
-      @noindex = 1
       @articles = Article.published_at(params.values_at(:year, :month, :day)).page(params[:page]).per(@limit)
     else
-      @noindex = 1 unless params[:page].blank?
       @articles = Article.published.page(params[:page]).per(@limit)
     end
 
-    @page_title = index_title
-    @description = index_description
-    @keywords = (this_blog.meta_keywords.empty?) ? "" : this_blog.meta_keywords
+    @page_title = this_blog.home_title_template
+    @description = this_blog.home_desc_template
+    if params[:year]
+      @page_title = this_blog.archives_title_template
+      @description = this_blog.archives_desc_template
+    elsif params[:page]
+      @page_title = this_blog.paginated_title_template
+      @description = this_blog.paginated_desc_template
+    end
+    @page_title = @page_title.to_title(@articles, this_blog, params)
+    @description = @description.to_title(@articles, this_blog, params)
+
+    @keywords = this_blog.meta_keywords
 
     suffix = (params[:page].nil? and params[:year].nil?) ? "" : "/"
 
@@ -106,7 +114,7 @@ class ArticlesController < ContentController
   def archives
     @articles = Article.find_published
     @page_title = this_blog.archives_title_template.to_title(@articles, this_blog, params)
-    @keywords = (this_blog.meta_keywords.empty?) ? "" : this_blog.meta_keywords
+    @keywords = this_blog.meta_keywords
     @description = this_blog.archives_desc_template.to_title(@articles, this_blog, params)
     @canonical_url = url_for(:only_path => false, :controller => 'articles', :action => 'archives')
   end
@@ -117,7 +125,7 @@ class ArticlesController < ContentController
       return
     end
 
-    set_headers
+    headers["Content-Type"] = "text/html; charset=utf-8"
     @comment = Comment.new(params[:comment])
     @controller = self
   end
@@ -130,11 +138,18 @@ class ArticlesController < ContentController
     redirect_to tags_path, :status => 301
   end
 
+  def preview_page
+    @page = Page.find(params[:id]) 
+    @canonical_url = ""
+    render 'view_page'
+  end
+  
+
   def view_page
     if(@page = Page.find_by_name(Array(params[:name]).map { |c| c }.join("/"))) && @page.published?
       @page_title = @page.title
-      @description = (this_blog.meta_description.empty?) ? "" : this_blog.meta_description
-      @keywords = (this_blog.meta_keywords.empty?) ? "" : this_blog.meta_keywords
+      @description = this_blog.meta_description
+      @keywords = this_blog.meta_keywords
       @canonical_url = @page.permalink_url
     else
       render "errors/404", :status => 404
@@ -163,7 +178,9 @@ class ArticlesController < ContentController
     @comment      = Comment.new
     @page_title   = this_blog.article_title_template.to_title(@article, this_blog, params)
     @description = this_blog.article_desc_template.to_title(@article, this_blog, params)
-    article_meta
+    groupings = @article.categories + @article.tags
+    @keywords = groupings.map { |g| g.name }.join(", ")
+    @canonical_url = @article.permalink_url
 
     auto_discovery_feed
     respond_to do |format|
@@ -174,13 +191,6 @@ class ArticlesController < ContentController
     end
   rescue ActiveRecord::RecordNotFound
     error("Post not found...")
-  end
-
-
-  def article_meta
-    groupings = @article.categories + @article.tags
-    @keywords = groupings.map { |g| g.name }.join(", ")
-    @canonical_url = @article.permalink_url
   end
 
   def render_articles_feed format
@@ -196,10 +206,6 @@ class ArticlesController < ContentController
     render "feedback_#{format}_feed", :layout => false
   end
 
-  def set_headers
-    headers["Content-Type"] = "text/html; charset=utf-8"
-  end
-
   def render_paginated_index(on_empty = _("No posts found..."))
     return error(on_empty, :status => 200) if @articles.empty?
     if this_blog.feedburner_url.empty?
@@ -209,36 +215,6 @@ class ArticlesController < ContentController
       @auto_discovery_url_atom = "http://feeds2.feedburner.com/#{this_blog.feedburner_url}"
     end
     render 'index'
-  end
-
-  def index_title
-    if params[:year]
-      return this_blog.archives_title_template.to_title(@articles, this_blog, params)
-    elsif params[:page]
-      return this_blog.paginated_title_template.to_title(@articles, this_blog, params)
-    else
-      this_blog.home_title_template.to_title(@articles, this_blog, params)
-    end
-  end
-
-  def index_description
-    if params[:year]
-      return this_blog.archives_desc_template.to_title(@articles, this_blog, params)
-    elsif params[:page]
-      return this_blog.paginated_desc_template.to_title(@articles, this_blog, params)
-    else
-      this_blog.home_desc_template.to_title(@articles, this_blog, params)
-    end
-  end
-
-  def time_delta(year, month = nil, day = nil)
-    from = Time.mktime(year, month || 1, day || 1)
-
-    to = from.next_year
-    to = from.next_month unless month.blank?
-    to = from + 1.day unless day.blank?
-    to = to - 1 # pull off 1 second so we don't overlap onto the next day
-    return from..to
   end
 
   def split_from_path path
