@@ -28,14 +28,33 @@ class SetupController < ApplicationController
 
     # FIXME: Crappy hack : by default, the auto generated post is user_id less and it makes Typo crash
     if User.count == 1
-      art = Article.find(:first)
-      art.user_id = @user.id
-      art.save
+      update_or_create_first_post_with_user @user
     end
+
     redirect_to :action => 'confirm'
   end
 
   private
+
+  # FIXME: Move to a setup concern that coordinates first blog, user, and post
+  def update_or_create_first_post_with_user user
+    art = Article.find(:first)
+    if art
+      art.user = user
+      art.save
+    else
+      Article.create(title: 'Hello World!',
+                     author: user.login,
+                     body: 'Welcome to Typo. This is your first article. Edit or delete it, then start blogging!',
+                     allow_comments: 1,
+                     allow_pings: 1,
+                     published: 1,
+                     permalink: 'hello-world',
+                     categories: [Category.find(:first)],
+                     user: user)
+    end
+  end
+
   def generate_password
     chars = ("a".."z").to_a + ("A".."Z").to_a + ("0".."9").to_a
     newpass = ""
