@@ -2,6 +2,14 @@ require 'set'
 require 'uri'
 
 class Content < ActiveRecord::Base
+  include Stateful
+
+  include ContentBase
+
+  # TODO: Move these calls to ContentBase
+  after_save :invalidates_cache?
+  after_destroy lambda { |c|  c.invalidates_cache?(true) }
+
   belongs_to :text_filter
 
   has_many :redirections
@@ -39,29 +47,13 @@ class Content < ActiveRecord::Base
 
   serialize :whiteboard
 
-  attr_accessor :just_changed_published_status
-  alias_method :just_changed_published_status?, :just_changed_published_status
-
-  after_save :invalidates_cache?
-  after_destroy lambda { |c|  c.invalidates_cache?(true) }
-
-  include Stateful
-
-  def invalidates_cache?(on_destruction = false)
-    @invalidates_cache ||= if on_destruction
-      just_changed_published_status? || published?
-    else
-      (changed? && published?) || just_changed_published_status?
-    end
-  end
-  
   def shorten_url
     return unless self.published
-    
+
     r = Redirect.new
     r.from_path = r.shorten
     r.to_path = self.permalink_url
-    
+
     # This because updating self.redirects.first raises ActiveRecord::ReadOnlyRecord
     unless (red = self.redirects.first).nil?
       return if red.to_path == self.permalink_url
@@ -74,20 +66,6 @@ class Content < ActiveRecord::Base
   end
 
   class << self
-    def content_fields *attribs
-      class_eval "def content_fields; #{attribs.inspect}; end"
-    end
-
-    def find_published(what = :all, options = {})
-      with_scope(:find => {:order => default_order, :conditions => {:published => true}}) do
-        find what, options
-      end
-    end
-
-    def default_order
-      'published_at DESC'
-    end
-
     def find_already_published(what = :all, at = nil, options = { })
       if what.respond_to?(:has_key?)
         what, options = :all, what
@@ -105,15 +83,14 @@ class Content < ActiveRecord::Base
 
       # Implement adapter-specific groupings below, or allow us to fall through to the generic ruby-side grouping
 
-      if defined?(ActiveRecord::ConnectionAdapters::MysqlAdapter) && self.connection.is_a?(ActiveRecord::ConnectionAdapters::MysqlAdapter)
+      if defined?(ActiveRecord::ConnectionAdapters::Mysql2Adapter) && self.connection.is_a?(ActiveRecord::ConnectionAdapters::Mysql2Adapter)
         # MySQL uses date_format
         find_by_sql("SELECT date_format(#{column_name}, '%Y-%m') AS publication #{from_where} GROUP BY publication ORDER BY publication DESC")
-
       elsif defined?(ActiveRecord::ConnectionAdapters::PostgreSQLAdapter) && self.connection.is_a?(ActiveRecord::ConnectionAdapters::PostgreSQLAdapter)
         # PostgreSQL uses to_char
         find_by_sql("SELECT to_char(#{column_name}, 'YYYY-MM') AS publication #{from_where} GROUP BY publication ORDER BY publication DESC")
 
-      else
+      else        
         # If we don't have an adapter-safe conversion from date -> YYYY-MM,
         # we'll do the GROUP BY server-side. There won't be very many objects
         # in this array anyway.
@@ -162,35 +139,6 @@ class Content < ActiveRecord::Base
     end
   end
 
-  def html_map field
-    content_fields.include? field
-  end
-
-  # Return HTML for some part of this object.
-  def html(field = :all)
-    if field == :all
-      generate_html(:all, content_fields.map{|f| self[f].to_s}.join("\n\n"))
-    elsif html_map(field)
-      generate_html(field)
-    else
-      raise "Unknown field: #{field.inspect} in content.html"
-    end
-  end
-
-  # Generate HTML for a specific field using the text_filter in use for this
-  # object.
-  def generate_html(field, text = nil)
-    text ||= self[field].to_s
-    html = text_filter.filter_text_for_content(blog, text, self) || text
-    html_postprocess(field,html).to_s
-  end
-
-  # Post-process the HTML.  This is a noop by default, but Comment overrides it
-  # to enforce HTML sanity.
-  def html_postprocess(field,html)
-    html
-  end
-
   def whiteboard
     self[:whiteboard] ||= Hash.new
   end
@@ -201,30 +149,6 @@ class Content < ActiveRecord::Base
     blog.text_filter_object
   end
 
-  # Grab the text filter for this object.  It's either the filter specified by
-  # self.text_filter_id, or the default specified in the default blog object.
-  def text_filter
-    if self[:text_filter_id] && !self[:text_filter_id].zero?
-      TextFilter.find(self[:text_filter_id])
-    else
-      default_text_filter
-    end
-  end
-
-  # Set the text filter for this object.
-  def text_filter= filter
-    self.text_filter_id = filter.to_text_filter.id
-  end
-
-  def blog
-    @blog ||= Blog.default
-  end
-
-  def publish!
-    self.published = true
-    self.save!
-  end
-
   def withdraw!
     self.withdraw
     self.save!
@@ -232,17 +156,6 @@ class Content < ActiveRecord::Base
 
   def published_at
     self[:published_at] || self[:created_at]
-  end
-
-  def send_notification_to_user(user)
-    notify_user_via_email(user)
-  end
-
-  def really_send_notifications
-    interested_users.each do |value|
-      send_notification_to_user(value)
-    end
-    return true
   end
 
   def get_rss_description
@@ -264,7 +177,7 @@ class Content < ActiveRecord::Base
 
   def short_url
     # Double check because of crappy data in my own old database
-    return unless self.published and self.redirects.count > 0
+    return unless self.published and self.redirects.size > 0
     blog.url_for(redirects.last.from_path, :only_path => false)
   end
 
