@@ -1,8 +1,18 @@
 require_dependency 'spam_protection'
-class Feedback < Content
-  set_table_name "feedback"
+class Feedback < ActiveRecord::Base
+  self.table_name = "feedback"
+
+  belongs_to :text_filter
 
   include TypoGuid
+
+  ## For fixing this not being Content anymore
+  include Stateful
+
+  include ContentBase
+
+  after_save :invalidates_cache?
+  after_destroy lambda { |c|  c.invalidates_cache?(true) }
 
   validate :feedback_not_closed, :on => :create
 
@@ -28,10 +38,6 @@ class Feedback < Content
 
   def self.default_order
     'created_at ASC'
-  end
-
-  def to_param
-    guid
   end
 
   def parent
@@ -123,14 +129,17 @@ class Feedback < Content
     end
   end
 
-  def mark_as_ham!
-    mark_as_ham
+  def change_state!
+    result = ''
+    if state.spam?
+      mark_as_ham
+      result = 'ham'
+    else
+      mark_as_spam
+      result = 'spam'
+    end
     save!
-  end
-
-  def mark_as_spam!
-    mark_as_spam
-    save
+    result
   end
 
   def report_as_spam
