@@ -8,19 +8,20 @@ class Theme
     @name, @path = name, path
   end
 
-  # TODO: Remove check for old-fashioned theme layout.
   def layout(action=:default)
     if action.to_s == 'view_page'
       if File.exists? "#{::Rails.root.to_s}/themes/#{name}/views/layouts/pages.html.erb"
-        return "layouts/pages.html.erb"
+        return "layouts/pages.html"
       end
+      # FIXME: This old-fashioned location will be removed soon
       if File.exists? "#{::Rails.root.to_s}/themes/#{name}/layouts/pages.html.erb"
-        return "#{::Rails.root.to_s}/themes/#{name}/layouts/pages.html.erb"
+        return "#{::Rails.root.to_s}/themes/#{name}/layouts/pages.html"
       end
     end
     if File.exists? "#{::Rails.root.to_s}/themes/#{name}/views/layouts/default.html.erb"
-      return "layouts/default.html.erb"
+      return "layouts/default.html"
     end
+    # FIXME: This old-fashioned location will be removed soon
     "#{::Rails.root.to_s}/themes/#{name}/layouts/default.html.erb"
   end
 
