@@ -17,8 +17,7 @@ module TypoBlog
 
     # I need the localization plugin to load first
     # Otherwise, I can't localize plugins <= localization
-    # Forcing manually the load of the textfilters plugins fixes the bugs with apache in production.
-    config.plugins = [ :localization, :prototype_legacy_helper, :all ]
+    config.plugins = [ :localization, :all ]
 
     config.autoload_paths += %W(
       app/apis
@@ -27,16 +26,11 @@ module TypoBlog
     # Activate observers that should always be running
     config.active_record.observers = :email_notifier, :web_notifier
 
-    # Turn of timestamped migrations
-    config.active_record.timestamped_migrations = false
+    # Turn om timestamped migrations
+    config.active_record.timestamped_migrations = true
 
     # Filter sensitive parameters from the log file
     config.filter_parameters << :password
-  end
-
-  if RUBY_VERSION < "1.9"
-    $KCODE = 'u'
-    require 'jcode'
   end
 
   # Load included libraries.
