@@ -489,7 +489,7 @@ Localization.define("es_MX") do |l|
   l.store "User's status", ""
   l.store "Active", ""
   l.store "Inactive", ""
-  l.store "Profile Settings", ""
+  l.store "Profile settings", ""
   l.store "Firstname", ""
   l.store "Lastname", ""
   l.store "Nickname", ""
@@ -499,9 +499,9 @@ Localization.define("es_MX") do |l|
   l.store "Send notification messages via email", "Enviar notificaciones por email"
   l.store "Send notification messages when new articles are posted", "Enviar notificaciones cuando un nuevo art&iacute;culo sea publicado"
   l.store "Send notification messages when comments are posted", "Enviar notificaciones cuando un nuevo comentario sea publicado"
-  l.store "Contact Options", ""
+  l.store "Contact options", ""
   l.store "Your site", ""
-  l.store "display url on public profile", ""
+  l.store "display URL on public profile", ""
   l.store "Your MSN", ""
   l.store "display MSN ID on public profile", ""
   l.store "Your Yahoo ID", ""
@@ -511,7 +511,7 @@ Localization.define("es_MX") do |l|
   l.store "Your AIM id", ""
   l.store "display AIM ID on public profile", ""
   l.store "Your Twitter username", ""
-  l.store "display twitter on public profile", ""
+  l.store "display Twitter on public profile", ""
   l.store "Tell us more about you", ""
 
   # app/views/admin/users/destroy.html.erb
@@ -594,7 +594,7 @@ Localization.define("es_MX") do |l|
   l.store "Welcome to your %s blog setup. Just fill in your blog title and your email, and Typo will take care of everything else", ""
 
   # app/views/shared/_confirm.html.erb
-  l.store "Congratulation!", ""
+  l.store "Congratulations!", ""
   l.store "You have successfully signed up", ""
   l.store "<strong>Login:</strong> %s", ""
   l.store "<strong>Password:</strong> %s", ""
