@@ -542,7 +542,7 @@ Localization.define("fr_FR") do |l|
   l.store "User's status", "Statut de l'utilisateur"
   l.store "Active", "Actif"
   l.store "Inactive", "Inactif"
-  l.store "Profile Settings", "Paramètres du profil"
+  l.store "Profile settings", "Paramètres du profil"
   l.store "Firstname", "Prénom"
   l.store "Lastname", "Nom"
   l.store "Nickname", "Surnom"
@@ -553,9 +553,9 @@ Localization.define("fr_FR") do |l|
   l.store "Send notification messages via email", "Envoi de notification des messages par email"
   l.store "Send notification messages when new articles are posted", "Envoi de notification de messages quand de nouveaux articles sont postés"
   l.store "Send notification messages when comments are posted", "Envoi de notification de messages quand des commentaires sont postés"
-  l.store "Contact Options", "Paramètres de contact"
+  l.store "Contact options", "Paramètres de contact"
   l.store "Your site", "Votre site"
-  l.store "display url on public profile", "Afficher votre site sur votre profil public"
+  l.store "display URL on public profile", "Afficher votre site sur votre profil public"
   l.store "Your MSN", "Adresse MSN"
   l.store "display MSN ID on public profile", "Afficher votre adresse MSN sur votre profil public"
   l.store "Your Yahoo ID", "Identifiant Yahoo"
@@ -565,7 +565,7 @@ Localization.define("fr_FR") do |l|
   l.store "Your AIM id", "Identifiant AIM"
   l.store "display AIM ID on public profile", "Afficher votre identifiant AIM sur votre profil public"
   l.store "Your Twitter username", "Identifiant Twitter"
-  l.store "display twitter on public profile", "Afficher votre identifiant Twitter sur votre profil public"
+  l.store "display Twitter on public profile", "Afficher votre identifiant Twitter sur votre profil public"
   l.store "Tell us more about you", "Dites nous en plus à votre sujet"
 
   # app/views/admin/users/edit.html.erb
@@ -652,7 +652,7 @@ Localization.define("fr_FR") do |l|
   l.store "Welcome to your %s blog setup. Just fill in your blog title and your email, and Typo will take care of everything else", ""
 
   # app/views/shared/_confirm.html.erb
-  l.store "Congratulation!", "Félicitations !"
+  l.store "Congratulations!", "Félicitations !"
   l.store "You have successfully signed up", "Vous vous êtes inscrit avec succès"
   l.store "<strong>Login:</strong> %s", "<strong>Identifiant&nbsp;:</strong> %s"
   l.store "<strong>Password:</strong> %s", "<strong>Mot de passe&nbsp;:</strong> %s"
