@@ -573,7 +573,7 @@ Localization.define("nb_NO") do |l|
   l.store "User's status", "Brukerens status"
   l.store "Active", "Aktiv"
   l.store "Inactive", "Inaktiv"
-  l.store "Profile Settings", "Profilinnstillinger"
+  l.store "Profile settings", "Profilinnstillinger"
   l.store "Firstname", "Fornavn"
   l.store "Lastname", "Etternavn"
   l.store "Nickname", "Nick"
@@ -584,9 +584,9 @@ Localization.define("nb_NO") do |l|
   l.store "Send notification messages via email", "Send beskjeder via epost"
   l.store "Send notification messages when new articles are posted", "Send beskjeder når nye artikler blir publisert"
   l.store "Send notification messages when comments are posted", "Send beskjeder når det kommer nye kommentarer"
-  l.store "Contact Options", "Kontaktinformasjon"
+  l.store "Contact options", "Kontaktinformasjon"
   l.store "Your site", "Din hjemmeside"
-  l.store "display url on public profile", "Vis hjemmeside på din profil"
+  l.store "display URL on public profile", "Vis hjemmeside på din profil"
   l.store "Your MSN", "Din MSN ID"
   l.store "display MSN ID on public profile", "Vis din MSN ID på din profil"
   l.store "Your Yahoo ID", "Din Yahoo ID"
@@ -596,7 +596,7 @@ Localization.define("nb_NO") do |l|
   l.store "Your AIM id", "Dit AIM ID"
   l.store "display AIM ID on public profile", "Vis dit AIM ID på din profil"
   l.store "Your Twitter username", "Ditt Twitter-brukernavn"
-  l.store "display twitter on public profile", "Vis Twitter på din profil"
+  l.store "display Twitter on public profile", "Vis Twitter på din profil"
   l.store "Tell us more about you", "Fortell litt mer om deg selv"
 
   # app/views/admin/users/destroy.html.erb
@@ -681,7 +681,7 @@ Localization.define("nb_NO") do |l|
   l.store "Welcome to your %s blog setup. Just fill in your blog title and your email, and Typo will take care of everything else", "Velkommen til din %s bloggoppsett. Bare fyll inn bloggtittel og epostadressen din, så vil Typo ta seg av resten"
 
   # app/views/shared/_confirm.html.erb
-  l.store "Congratulation!", "Gratulerer!"
+  l.store "Congratulations!", "Gratulerer!"
   l.store "You have successfully signed up", "Du er nå registrert"
   l.store "<strong>Login:</strong> %s", "<strong>Logg inn:</strong> %s"
   l.store "<strong>Password:</strong> %s", "<strong>Passord:</strong> %s"
