@@ -43,7 +43,7 @@
 #   the scope of the migration with a number based on the migration
 #   number, say Bare3Article
 #
-#   Use set_table_name if the object doesn't calculate its table name
+#   Use table_name= if the object doesn't calculate its table name
 #   properly.
 #
 # - Then, use the bare objects to migrate the data:
@@ -61,8 +61,8 @@ module BareMigration
     base.extend(ClassMethods)
 
     # Set the table name by eradicating "Bare" from the calculated table name.
-    # You can still use set_table_name if this is wrong.
-    base.set_table_name(base.table_name.sub(/.*?bare\d*_?/, ''))
+    # You can still use table_name= if this is wrong.
+    base.table_name = base.table_name.sub(/.*?bare\d*_?/, '')
 
     # Default to ignoring the inheritance column
     base.inheritance_column = :ignore_inheritance_column
