@@ -10,12 +10,21 @@ module Migrator
     Dir["#{migrations_path}/[0-9]*_*.rb"].sort_by { |name| name.scan(/\d+/).first.to_i }
   end
 
+  def self.needed_migrations(from)
+    migs = []
+    available_migrations.each do |mig|
+      next if mig.gsub(migrations_path,"").split('_').first.tr('^0-9','').to_i <= from
+      migs.push(mig)
+    end
+    migs
+  end
+
   def self.current_schema_version
     ActiveRecord::Migrator.current_version rescue 0
   end
 
   def self.max_schema_version
-    available_migrations.size
+    available_migrations.last.gsub(migrations_path,"").split('_').first.tr('^0-9','').to_i
   end
 
   def self.db_supports_migrations?
