@@ -2,23 +2,9 @@
 require 'spec_helper'
 
 describe ArticlesController do
-  render_views
-
   before(:each) do
-    #TODO Need to reduce user, but allow to remove user fixture...
-    FactoryGirl.create(:user,
-            :login => 'henri',
-            :password => 'whatever',
-            :name => 'Henri',
-            :email => 'henri@example.com',
-            :settings => {:notify_watch_my_articles => false, :editor => 'simple'},
-            :text_filter => FactoryGirl.create(:markdown),
-            :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN),
-            :notify_via_email => false,
-            :notify_on_new_articles => false,
-            :notify_on_comments => false,
-            :state => 'active')
-    FactoryGirl.create(:blog, :custom_tracking_field => '<script src="foo.js" type="text/javascript"></script>') 
+    build_stubbed :blog
+    create :user
   end
 
 
@@ -46,24 +32,23 @@ describe ArticlesController do
       assigns[:articles].should_not be_empty
     end
 
-    it 'should have good link feed rss' do
-      response.should have_selector('head>link[href="http://test.host/articles.rss"]')
-    end
+    context "with the view rendered" do
+      render_views
+      it 'should have good link feed rss' do
+        response.should have_selector('head>link[href="http://test.host/articles.rss"]')
+      end
 
-    it 'should have good link feed atom' do
-      response.should have_selector('head>link[href="http://test.host/articles.atom"]')
-    end
+      it 'should have good link feed atom' do
+        response.should have_selector('head>link[href="http://test.host/articles.atom"]')
+      end
 
-    it 'should have a canonical url' do
-      response.should have_selector('head>link[href="http://test.host/"]')
-    end
-    
-    it 'should have googd title' do 
-      response.should have_selector('title', :content => "test blog | test subtitles")
-    end
-    
-    it 'should have a custom tracking field' do      
-      response.should have_selector('head>script[src="foo.js"]')
+      it 'should have a canonical url' do
+        response.should have_selector('head>link[href="http://test.host/"]')
+      end
+
+      it 'should have googd title' do
+        response.should have_selector('title', :content => "test blog | test subtitles")
+      end
     end
   end
 
@@ -89,30 +74,29 @@ describe ArticlesController do
         assigns[:articles].should_not be_nil
       end
 
-      it 'should have good feed rss link' do
-        response.should have_selector('head>link[href="http://test.host/search/a.rss"]')
-      end
-
-      it 'should have good feed atom link' do
-        response.should have_selector('head>link[href="http://test.host/search/a.atom"]')
-      end
-
-      it 'should have a canonical url' do
-        response.should have_selector('head>link[href="http://test.host/search/a"]')
-      end
-      
-      it 'should have a good title' do
-        response.should have_selector('title', :content => "Results for a | test blog")
-      end
-
-      it 'should have content markdown interpret and without html tag' do
-        response.should have_selector('div') do |div|
-          div.should contain(/in markdown format\n\n\nwe\nuse\nok to define a link\n\n...\n/)
+      context "with the view rendered" do
+        render_views
+        it 'should have good feed rss link' do
+          response.should have_selector('head>link[href="http://test.host/search/a.rss"]')
         end
-      end
 
-      it 'should have a custom tracking field' do      
-        response.should have_selector('head>script[src="foo.js"]')
+        it 'should have good feed atom link' do
+          response.should have_selector('head>link[href="http://test.host/search/a.atom"]')
+        end
+
+        it 'should have a canonical url' do
+          response.should have_selector('head>link[href="http://test.host/search/a"]')
+        end
+
+        it 'should have a good title' do
+          response.should have_selector('title', :content => "Results for a | test blog")
+        end
+
+        it 'should have content markdown interpret and without html tag' do
+          response.should have_selector('div') do |div|
+            div.should contain(/in markdown format\n\n\nwe\nuse\nok to define a link\n\n...\n/)
+          end
+        end
       end
     end
 
@@ -121,7 +105,6 @@ describe ArticlesController do
       response.should be_success
       response.should render_template('index_rss_feed')
       @layouts.keys.compact.should be_empty
-      response.should_not have_selector('head>script[src="foo.js"]')    
     end
 
     it 'should render feed atom by search' do
@@ -129,7 +112,6 @@ describe ArticlesController do
       response.should be_success
       response.should render_template('index_atom_feed')
       @layouts.keys.compact.should be_empty
-      response.should_not have_selector('head>script[src="foo.js"]')      
     end
 
     it 'search with empty result' do
@@ -137,7 +119,7 @@ describe ArticlesController do
       response.should render_template('articles/error')
       assigns[:articles].should be_empty
     end
-    
+
   end
 
   describe '#livesearch action' do
@@ -164,8 +146,11 @@ describe ArticlesController do
         response.should render_template('live_search')
       end
 
-      it 'should not have h3 tag' do
-        response.should have_selector("h3")
+      context "with the view rendered" do
+        render_views
+        it 'should not have h3 tag' do
+          response.should have_selector("h3")
+        end
       end
 
       it "should assign @search the search string" do
@@ -176,16 +161,18 @@ describe ArticlesController do
   end
 
 
-  it 'archives' do
-    3.times { FactoryGirl.create(:article) }
-    get 'archives'
-    response.should render_template(:archives)
-    assigns[:articles].should_not be_nil
-    assigns[:articles].should_not be_empty
+  describe '#archives' do
+    render_views
+    it "works" do
+      3.times { FactoryGirl.create(:article) }
+      get 'archives'
+      response.should render_template(:archives)
+      assigns[:articles].should_not be_nil
+      assigns[:articles].should_not be_empty
 
-    response.should have_selector('head>link[href="http://test.host/archives"]')
-    response.should have_selector('title', :content => "Archives for test blog")
-    response.should have_selector('head>script[src="foo.js"]')
+      response.should have_selector('head>link[href="http://test.host/archives"]')
+      response.should have_selector('title', :content => "Archives for test blog")
+    end
   end
 
   describe 'index for a month' do
@@ -204,16 +191,15 @@ describe ArticlesController do
       assigns[:articles].should_not be_empty
     end
 
-    it 'should have a canonical url' do
-      response.should have_selector('head>link[href="http://test.host/2004/4/"]')
-    end
-    
-    it 'should have a good title' do
-      response.should have_selector('title', :content => "Archives for test blog")
-    end
-    
-    it 'should have a custom tracking field' do      
-      response.should have_selector('head>script[src="foo.js"]')
+    context "with the view rendered" do
+      render_views
+      it 'should have a canonical url' do
+        response.should have_selector('head>link[href="http://test.host/2004/4/"]')
+      end
+
+      it 'should have a good title' do
+        response.should have_selector('title', :content => "Archives for test blog")
+      end
     end
   end
 
@@ -221,7 +207,6 @@ end
 
 describe ArticlesController, "nosettings" do
   before(:each) do
-    Blog.delete_all
     @blog = Blog.new.save
   end
 
@@ -234,11 +219,7 @@ end
 
 describe ArticlesController, "nousers" do
   before(:each) do
-    FactoryGirl.create(:blog)
-    User.stub!(:count).and_return(0)
-    @user = mock("user")
-    @user.stub!(:reload).and_return(@user)
-    User.stub!(:new).and_return(@user)
+    build_stubbed(:blog)
   end
 
   it 'redirects to signup' do
@@ -249,7 +230,7 @@ end
 
 describe ArticlesController, "feeds" do
   before(:each) do
-    FactoryGirl.create(:blog)
+    build_stubbed(:blog)
     @article1 = FactoryGirl.create(:article,
                                :created_at => Time.now - 1.day)
     FactoryGirl.create(:trackback, :article => @article1, :published_at => Time.now - 1.day,
@@ -294,7 +275,7 @@ end
 
 describe ArticlesController, "the index" do
   before(:each) do
-    FactoryGirl.create(:blog) 
+    build_stubbed(:blog)
     FactoryGirl.create(:user, :login => 'henri', :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN))
   end
 
@@ -307,7 +288,7 @@ end
 
 describe ArticlesController, "previewing" do
   render_views
-  before(:each) { @blog = FactoryGirl.create(:blog) }
+  before(:each) { @blog = build_stubbed(:blog) }
 
   describe 'with non logged user' do
     before :each do
@@ -354,42 +335,18 @@ describe ArticlesController, "redirecting" do
 
   describe "with explicit redirects" do
     it 'should redirect from known URL' do
-      #TODO Need to reduce user, but allow to remove user fixture...
-      FactoryGirl.create(:user,
-              :login => 'henri',
-              :password => 'whatever',
-              :name => 'Henri',
-              :email => 'henri@example.com',
-              :settings => {:notify_watch_my_articles => false, :editor => 'simple'},
-              :text_filter => FactoryGirl.create(:markdown),
-              :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN),
-              :notify_via_email => false,
-              :notify_on_new_articles => false,
-              :notify_on_comments => false,
-              :state => 'active')
-      FactoryGirl.create(:blog)
-      FactoryGirl.create(:redirect)
+      build_stubbed(:blog)
+      create(:user)
+      create(:redirect)
       get :redirect, :from => "foo/bar"
       assert_response 301
       response.should redirect_to("http://test.host/someplace/else")
     end
 
     it 'should not redirect from unknown URL' do
-      #TODO Need to reduce user, but allow to remove user fixture...
-      FactoryGirl.create(:user,
-              :login => 'henri',
-              :password => 'whatever',
-              :name => 'Henri',
-              :email => 'henri@example.com',
-              :settings => {:notify_watch_my_articles => false, :editor => 'simple'},
-              :text_filter => FactoryGirl.create(:markdown),
-              :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN),
-              :notify_via_email => false,
-              :notify_on_new_articles => false,
-              :notify_on_comments => false,
-              :state => 'active')
-      FactoryGirl.create(:blog)
-      FactoryGirl.create(:redirect)
+      build_stubbed(:blog)
+      create(:user)
+      create(:redirect)
       get :redirect, :from => "something/that/isnt/there"
       assert_response 404
     end
@@ -401,23 +358,8 @@ describe ArticlesController, "redirecting" do
     # redirects?
     describe 'and non-empty relative_url_root' do
       before do
-        b = FactoryGirl.create(:blog, :base_url => "http://test.host/blog")
-        #TODO Need to reduce user, but allow to remove user fixture...
-        FactoryGirl.create(:user,
-                :login => 'henri',
-                :password => 'whatever',
-                :name => 'Henri',
-                :email => 'henri@example.com',
-                :settings => {:notify_watch_my_articles => false, :editor => 'simple'},
-                :text_filter => FactoryGirl.create(:markdown),
-                :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN),
-                :notify_via_email => false,
-                :notify_on_new_articles => false,
-                :notify_on_comments => false,
-                :state => 'active')
-
-        # XXX: The following has no effect anymore.
-        # request.env["SCRIPT_NAME"] = "/blog"
+        build_stubbed(:blog, :base_url => "http://test.host/blog")
+        create(:user)
       end
 
       it 'should redirect' do
@@ -444,7 +386,7 @@ describe ArticlesController, "redirecting" do
   end
 
   it 'should get good article with utf8 slug' do
-    FactoryGirl.create(:blog)
+    build_stubbed(:blog)
     utf8article = FactoryGirl.create(:utf8article, :permalink => 'ルビー',
                                  :published_at => Time.utc(2004, 6, 2))
     get :redirect, :from => '2004/06/02/ルビー'
@@ -453,7 +395,7 @@ describe ArticlesController, "redirecting" do
 
   # NOTE: This is needed because Rails over-unescapes glob parameters.
   it 'should get good article with pre-escaped utf8 slug using unescaped slug' do
-    FactoryGirl.create(:blog)
+    build_stubbed(:blog)
     utf8article = FactoryGirl.create(:utf8article, :permalink => '%E3%83%AB%E3%83%93%E3%83%BC',
                                  :published_at => Time.utc(2004, 6, 2))
     get :redirect, :from => '2004/06/02/ルビー'
@@ -462,7 +404,7 @@ describe ArticlesController, "redirecting" do
 
   describe 'accessing old-style URL with "articles" as the first part' do
     it 'should redirect to article' do
-      FactoryGirl.create(:blog)
+      build_stubbed(:blog)
       article = FactoryGirl.create(:article, :permalink => 'second-blog-article',
                         :published_at => '2004-04-01 02:00:00',
                         :updated_at => '2004-04-01 02:00:00',
@@ -473,7 +415,7 @@ describe ArticlesController, "redirecting" do
     end
 
     it 'should redirect to article with url_root' do
-      b = FactoryGirl.create(:blog, :base_url => "http://test.host/blog")
+      b = build_stubbed(:blog, :base_url => "http://test.host/blog")
       article = FactoryGirl.create(:article, :permalink => 'second-blog-article',
                         :published_at => '2004-04-01 02:00:00',
                         :updated_at => '2004-04-01 02:00:00',
@@ -484,7 +426,7 @@ describe ArticlesController, "redirecting" do
     end
 
     it 'should redirect to article with articles in url_root' do
-      b = FactoryGirl.create(:blog, :base_url => "http://test.host/aaa/articles/bbb")
+      b = build_stubbed(:blog, :base_url => "http://test.host/aaa/articles/bbb")
       article = FactoryGirl.create(:article, :permalink => 'second-blog-article',
                         :published_at => '2004-04-01 02:00:00',
                         :updated_at => '2004-04-01 02:00:00',
@@ -498,7 +440,7 @@ describe ArticlesController, "redirecting" do
   describe 'with permalink_format like %title%.html' do
 
     before(:each) do
-      b = FactoryGirl.create(:blog, :permalink_format => '/%title%.html')
+      b = build_stubbed(:blog, :permalink_format => '/%title%.html')
 
       @article = FactoryGirl.create(:article, :permalink => 'second-blog-article',
                          :published_at => '2004-04-01 02:00:00',
@@ -604,7 +546,7 @@ describe ArticlesController, "redirecting" do
 
   describe "with a format containing a fixed component" do
     before(:each) do
-      b = FactoryGirl.create(:blog, :permalink_format => '/foo/%title%')
+      b = build_stubbed(:blog, :permalink_format => '/foo/%title%')
 
       @article = FactoryGirl.create(:article)
     end
@@ -622,7 +564,7 @@ describe ArticlesController, "redirecting" do
 
   describe "with a custom format with several fixed parts and several variables" do
     before(:each) do
-      b = FactoryGirl.create(:blog, :permalink_format => '/foo/bar/%year%/%month%/%title%')
+      b = build_stubbed(:blog, :permalink_format => '/foo/bar/%year%/%month%/%title%')
 
       @article = FactoryGirl.create(:article)
     end
@@ -657,7 +599,7 @@ describe ArticlesController, "password protected" do
   render_views
 
   before do
-    b = FactoryGirl.create(:blog, :permalink_format => '/%title%.html')
+    b = build_stubbed(:blog, :permalink_format => '/%title%.html')
     @article = FactoryGirl.create(:article, :password => 'password')
   end
 
@@ -681,26 +623,12 @@ end
 
 describe ArticlesController, "assigned keywords" do
   before do
-    @blog = FactoryGirl.create(:blog)
-    #TODO Need to reduce user, but allow to remove user fixture...
-    FactoryGirl.create(:user,
-            :login => 'henri',
-            :password => 'whatever',
-            :name => 'Henri',
-            :email => 'henri@example.com',
-            :settings => {:notify_watch_my_articles => false, :editor => 'simple'},
-            :text_filter => FactoryGirl.create(:markdown),
-            :profile => FactoryGirl.create(:profile_admin, :label => Profile::ADMIN),
-            :notify_via_email => false,
-            :notify_on_new_articles => false,
-            :notify_on_comments => false,
-            :state => 'active')
-
+    @blog = build_stubbed(:blog)
+    create :user
   end
 
   it 'article with categories should have meta keywords' do
     @blog.permalink_format = '/%title%.html'
-    @blog.save
     category = FactoryGirl.create(:category)
     article = FactoryGirl.create(:article, :categories => [category])
     get :redirect, :from => "#{article.permalink}.html"
@@ -709,7 +637,6 @@ describe ArticlesController, "assigned keywords" do
 
   it 'article with neither categories nor tags should not have meta keywords' do
     @blog.permalink_format = '/%title%.html'
-    @blog.save
     article = FactoryGirl.create(:article)
     get :redirect, :from => "#{article.permalink}.html"
     assigns(:keywords).should == ""
@@ -722,8 +649,44 @@ describe ArticlesController, "assigned keywords" do
 
   it 'index without option but with blog keywords should have meta keywords' do
     @blog.meta_keywords = "typo, is, amazing"
-    @blog.save
     get 'index'
     assigns(:keywords).should == "typo, is, amazing"
+  end
+end
+
+describe ArticlesController, "preview page" do
+  render_views
+  before(:each) { @blog = build_stubbed(:blog) }
+
+  describe 'with non logged user' do
+    before :each do
+      @request.session = {}
+      get :preview, :id => FactoryGirl.create(:article).id
+    end
+
+    it 'should redirect to login' do
+      response.should redirect_to(:controller => "accounts", :action => "login")
+    end
+  end
+
+  describe 'with logged user' do
+    before :each do
+      henri = create(:user, :login => 'henri', :profile => create(:profile_admin, :label => Profile::ADMIN))
+      @request.session = { :user => henri.id }
+      @page = FactoryGirl.create(:page)
+    end
+
+    with_each_theme do |theme, view_path|
+      it "should render template #{view_path}/articles/view_page" do
+        @blog.theme = theme if theme
+        get :preview_page, :id => @page.id
+        response.should render_template('articles/view_page')
+      end
+    end
+
+    it 'should assigns article define with id' do
+      get :preview_page, :id => @page.id
+      assigns[:page].should == @page
+    end
   end
 end
