@@ -16,7 +16,9 @@ describe TagsController, "/index" do
     specify { response.should be_success }
     specify { response.should render_template('articles/groupings') }
     specify { assigns(:groupings).should_not be_empty }
-    specify { response.body.should have_selector('ul.tags[id="taglist"]') }
+    it "has a list of tags" do
+      response.body.should have_selector('ul.tags[id="taglist"]')
+    end
   end
 
   describe "if :index template exists" do
@@ -151,58 +153,59 @@ end
 
 describe TagsController, "SEO Options" do
   render_views
-  
+
   before(:each) do 
     @blog = FactoryGirl.create(:blog)
     @a = FactoryGirl.create(:article)
     @foo = FactoryGirl.create(:tag, :name => 'foo', :articles => [@a])
   end
-  
-  it 'should have rel nofollow' do
-    @blog.unindex_tags = true
-    @blog.save
-    
-    get 'show', :id => 'foo'
-    response.should have_selector('head>meta[content="noindex, follow"]')
-  end
 
-  it 'should not have rel nofollow' do
-    @blog.unindex_tags = false
-    @blog.save
-    
-    get 'show', :id => 'foo'
-    response.should_not have_selector('head>meta[content="noindex, follow"]')
-  end
-  # meta_keywords
-  
-  it 'should not have meta keywords with deactivated option and no blog keywords' do
-    @blog.use_meta_keyword = false
-    @blog.save
-    get 'show', :id => 'foo'
-    response.should_not have_selector('head>meta[name="keywords"]')
-  end
+  describe "rendering list article with a given tag" do
+    it 'contains rel nofollow in head when blog is configure to unindex_tags' do
+      @blog.unindex_tags = true
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should have_selector('head>meta[content="noindex, follow"]')
+    end
 
-  it 'should not have meta keywords with deactivated option and blog keywords' do
-    @blog.use_meta_keyword = false
-    @blog.meta_keywords = "foo, bar, some, keyword"
-    @blog.save
-    get 'show', :id => 'foo'
-    response.should_not have_selector('head>meta[name="keywords"]')
-  end
+    it 'not contains rel nofollow when blog is configure to index_tags' do
+      @blog.unindex_tags = false
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should_not have_selector('head>meta[content="noindex, follow"]')
+    end
 
-  it 'should not have meta keywords with activated option and no blog keywords' do
-    @blog.use_meta_keyword = true
-    @blog.save
-    get 'show', :id => 'foo'
-    response.should_not have_selector('head>meta[name="keywords"]')
-  end
+    it 'not contains meta keywords with deactivated option and no blog keywords' do
+      @blog.use_meta_keyword = false
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should_not have_selector('head>meta[name="keywords"]')
+    end
 
-  it 'should have meta keywords with activated option and blog keywords' do
-    @blog.use_meta_keyword = true
-    @blog.meta_keywords = "foo, bar, some, keyword"
-    @blog.save
-    get 'show', :id => 'foo'
-    response.should have_selector('head>meta[name="keywords"]')
+
+    it 'should not have meta keywords with deactivated option and blog keywords' do
+      @blog.use_meta_keyword = false
+      @blog.meta_keywords = "foo, bar, some, keyword"
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should_not have_selector('head>meta[name="keywords"]')
+    end
+
+    it 'should not have meta keywords with activated option and no blog keywords' do
+      @blog.use_meta_keyword = true
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should_not have_selector('head>meta[name="keywords"]')
+    end
+
+    it 'should have meta keywords with activated option and blog keywords' do
+      @blog.use_meta_keyword = true
+      @blog.meta_keywords = "foo, bar, some, keyword"
+      @blog.save
+      get 'show', :id => 'foo'
+      response.should have_selector('head>meta[name="keywords"]')
+    end
+
   end
 
 end
