@@ -123,8 +123,8 @@ describe "articles/index_rss_feed.rss.builder" do
         render
       end
 
-      it "shows the body and extended content in the feed" do
-        rendered_entry.css("description").first.content.should =~ /public info.*and more/m
+      it "shows the body content in the feed" do
+        rendered_entry.css("description").first.content.should =~ /public info/
       end
 
       it "shows the RSS description in the feed" do
